調整資料模型，將 bookingId 和 id 欄位類型改為 Long，更新相關查詢及測試，並修正 application.properties 中的資料庫連接設定

+++ src/main/java/tw/rating/ratingproject/controller/ChatMessageRequest.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ChatMessageRequest {
-    private Integer bookingId;
+    private Long bookingId;
     private Integer senderId;
     private String content;
 }
